fix: address PR comments + refactors
Remove the return statement from saveFile.
Align getItemAttributes with the rest of the protocol's methods: to receive a URL instead of a plain-text path.
Add the withSecurityScopedAccess verification to all operations.
Add read file validations to validate the behaviour when the file doesn't exists.
Refactor the code a bit, to remove unnecessary protocol logic.